BSP-41: Add OpenAPI Specifications
- Add doc generator
- collect docs via NATS
