+++ src/index.ts
@@ -1,13 +1,10 @@
-import { connect, ErrorCode, JSONCodec } from "nats";
+import { connect } from "nats";
 import express, { json } from "express";
 
-const jsonCodec = JSONCodec();
+import restToNatsBridge from "./natsToRestBridge";
+import serveOpenAPIDocs from "./openAPIDocs";
 
 const { NATS_URL = "nats://localhost:4222", HTTP_PORT = 8000 } = process.env;
-
-interface PlatformResponse extends Object {
-    error?: string;
-}
 
 async function init() {
     const natsConnection = await connect({
@@ -18,45 +15,8 @@ async function init() {
 
     app.use(json());
 
-    app.use(async function restToNatsBridge(
-        req: express.Request,
-        res: express.Response,
-        next: express.NextFunction
-    ) {
-        const { method, path, body, query } = req;
-        const subjectName = `${method}:${path.substring(1).split("/").join(".")}`;
-
-        console.log(`[AIRLOCK] Request on ${subjectName}`);
-
-        try {
-            const reply = await natsConnection.request(
-                subjectName,
-                jsonCodec.encode({
-                    body, query
-                }),
-                { timeout: 30000 }
-            );
-
-            const response = jsonCodec.decode(reply.data) as PlatformResponse;
-
-            if (response.error) {
-                res.status(400).send({
-                    message: response.error
-                });
-            } else {
-                res.status(200).send(reply.data);
-            }
-        } catch (err) {
-            if (err.code === ErrorCode.NoResponders) {
-                res.status(404).send();
-            } else {
-                console.error(err);
-                res.status(500).send();
-            }
-        }
-
-        next();
-    });
+    app.use(serveOpenAPIDocs);
+    app.use(restToNatsBridge(natsConnection));
 
     console.info(`[AIRLOCK] Connected to Nats ${natsConnection.getServer()}`);
